Move tasks config to somewhere easier to find
