+++ tasks/rtl/build.js
@@ -27,7 +27,7 @@ var
   config       = require('../config/user'),
 
   // install config
-  tasks        = require('../config/project/tasks'),
+  tasks        = require('../config/tasks'),
   install      = require('../config/project/install'),
 
   // shorthand
