+++ tasks/rtl/watch.js
@@ -29,7 +29,7 @@ var
   config       = require('../config/user'),
 
   // task config
-  tasks        = require('../config/project/tasks'),
+  tasks        = require('../config/tasks'),
   install      = require('../config/project/install'),
 
   // shorthand
